rename: core -> common으로 폴더명 변경과 import ScheduleModal 경로 변경 (#102)
